requirements.txt
absl-py==2.3.1
accelerate==1.8.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
apturl==0.5.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
basicsr==1.4.2
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
carvekit-colab==4.1.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
clip==0.2.0
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
contourpy==1.3.2
crcmod==1.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==3.1.2
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
diffusers==0.34.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.14.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gradio==3.41.0
gradio_client==0.5.0
groovy==0.1.2
-e git+https://github.com/IDEA-Research/GroundingDINO.git@856dde20aee659246248e20734ef9ba5214f5e44#egg=groundingdino
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.3
imageio==2.37.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
insightface==0.7.3
invisible-watermark==0.2.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==23.5.0
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lightning-utilities==0.14.3
lmdb==1.7.2
loguru==0.7.3
louis==3.20.0
lpips==0.1.4
macaroonbakery==1.3.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcv==2.1.0
mmengine==0.10.7
mmsegmentation==1.2.2
model-index==0.1.11
modelscope==1.28.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.6.3
narwhals==1.47.0
natsort==8.4.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.4
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
open-clip-torch==2.24.0
openai==1.7.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.18
oss2==2.17.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
peft==0.15.2
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.3.8
portalocker==3.2.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycocotools==2.0.10
pycryptodome==3.23.0
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytorch-msssim==1.0.0
pytz==2023.4
PyWavelets==1.8.0
pyxdg==0.27
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.28.2
rich==13.4.2
rpds-py==0.26.0
ruff==0.12.3
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.7.0
scipy==1.15.3
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simsimd==6.5.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.1
stringzilla==3.12.5
supervision==0.25.1
sympy==1.13.3
systemd-python==234
tabulate==0.9.0
taming-transformers==0.0.1
tb-nightly==2.20.0a20250710
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.17
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchmetrics==1.7.4
torchsummary==1.5.1
torchvision==0.16.1+cu121
tornado==6.5.1
tqdm==4.65.2
traitlets==5.14.3
transformers==4.43.0
triton==2.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.35.0
wadllib==1.3.6
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
xdg==5
xformers==0.0.23
xkit==0.0.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zipp==1.0.0
